app2/requirements.txt
numpy
torch
Pillow
mtcnn
scikit-learn
opencv-python
face-recognition
facenet-pytorch
tensorflow